+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
-- Add $password_reset_requested event with $succeeded and $failed statuses
+- Add Log action for $password_reset_requested event with $succeeded and $failed statuses
+- Add Risk action for $profile_update event with no status and Log action for $profile_update event with $succeeded and $failed statuses
 
 ## [0.1.0] - 2021-07-08